Added data table & info
bulgeless data table in fits and ascii format, with table attributes as
screenshot (not sure how to export that as text from TOPCAT).

To re-make plot, stellar mass is LOGMSTAR_BALDRY06 and u-r should be
PETROMAG_U - PETROMAG_R. All other columns are options for plotting
other quantities (such as absolute magnitudes instead of mass or other
colours) or possible scale considerations for the toon.
L_second_over_L_host_tot is the "bulge"-to-total ratio, which is very
small for this sample. L_PS_over_L_Host is the r-band AGN-to-host
ratio. lL_bol_use is the AGN bolometric luminosity (in log, cgs). etc.